Add the CNN training and validation file
The training and validation file was created as a Jupyter Notebook. It is necessary to upload the dataset and a CSV file to run the training. The CSV file that must contain the files at the dataset and the and the corresponding class. As output, the notebook download the CNN parameters, the training and validation data, the image with validation and training data and the confusion matrix for all data provided by the dataset.