Add Fuzzy Inference System I and II
This file receives a dataset in .mat format. It converts the dataset to .npy format. The file contains the Fuzzy Inference System I and II. The Fuzzy Inference System I perform the  event classification, while the  Fuzzy Inference System II perform the fault classification. As output, the file provide a confusion matrix.